+++ src/compiler/transformers/map-imports-to-path-aliases.ts
@@ -60,7 +60,7 @@ export const mapImportsToPathAliases = (
               // from the import path
               const extensionRegex = new RegExp(
                 Object.values(ts.Extension)
-                  .map((extension) => `${extension}$`)
+                  .map((extension) => `${extension.replace('.', '\\.')}$`)
                   .join('|'),
               );