+++ src/internal/stencil-core/index.d.ts
@@ -41,6 +41,7 @@ export {
   Host,
   Listen,
   Method,
+  MixedInCtor,
   Mixin,
   MixinFactory,
   Prop,